==> 0.Code/1.Data_Cleaning.ipynb <==
Cleaning the data

import pandas as pd
import os

# Load your data
#r'C:/Users/gusta/Documents/KTH/TriOptima/trioptima/trioptima/'
#'/Users/elliotlindestam/Documents/Skola/Indek icloud/trioptima/'
your_path = '/Users/elliotlindestam/Documents/Skola/Indek icloud/trioptima/'
test_folder_path = your_path + '6.Active Data/Test Data/'
train_folder_path = your_path + '6.Active Data/Train Model Data/'

def clean_train(folder_path):
    try:
        # Get file in the folder
        files = os.listdir(folder_path)
        #MAC Issue
        files = [f for f in files if f != '.DS_Store']
        file = files[0]
        file_name = files[0][:-4]
        
        # Check if there are any files
        if len(files) == 0:
            print("No files found in the folder.")

        else:
            file_path = os.path.join(folder_path, file)
            data = pd.read_csv(file_path, dtype=str)
            print(file)
            
    except Exception as e:
        print(f"An error occurred: {e}")

    # Drop 'disseminationTimestamp' and its derived columns
    data = data.drop(columns=['disseminationTimestamp'])#,'sDRreceiptTimestamp','disseminationIdentifier'])

    # Convert any empty strings to NaN
    data = data.map(lambda x: None if x == '' else x)

    # Set your threshold
    threshold = 0.3 * len(data)

    # Collect removed columns and percentages
    removed_columns = []
    nan_percentages = []

    # Identify columns to be removed and collect their names and NaN percentages
    for column in data.columns:
        nan_count = data[column].isna().sum()
        nan_percentage = nan_count / len(data)
        
        if nan_percentage >= 0.3:
            removed_columns.append(column)
            nan_percentages.append(f"{nan_percentage:.2%}")

    # Display removed columns and percentages in a table
    removed_data = pd.DataFrame({
        'Removed Column': removed_columns,
        '% NaN Values': nan_percentages
    })
    print("Removed Columns:")
    
    # Create Interaction variables
    data['CurrencyIV'] = data.apply(lambda row: compare_item(row, 'leg1NotionalCurrency','leg2NotionalCurrency'), axis=1)
    data['NominalIV'] = data.apply(lambda row: compare_item(row, 'leg1NotionalAmount','leg2NotionalAmount'), axis=1)
    
        # Define the non-numeric values for each column
    non_numeric_values_leg1 = ['QTR', '6D']
    non_numeric_values_leg2 = ['ZC', 'QTR', '6WK', 'SA']

    # Identify rows with non-numeric values in either of the two columns
    non_numeric_rows = data[
        data['leg1FixedRatePaymentFrequencyMultiplier'].isin(non_numeric_values_leg1) |
        data['leg2UnderlierTenorMultiplier'].isin(non_numeric_values_leg2)
    ]

    # Remove these rows from the DataFrame
    data = data.drop(non_numeric_rows.index)
    data = data.dropna(subset=['leg1FixedRatePaymentFrequencyPeriod','leg2UnderlierTenorPeriod'
                               ,'leg1FixedRatePaymentFrequencyMultiplier', 'leg2UnderlierTenorMultiplier',
                               'leg2UnderlierID'])

    # Remove identified columns
    data = data.drop(columns=removed_columns)

    # Display remaining columns
    remaining_columns = list(data.columns)
    print("\nRemaining Columns:")
    print(remaining_columns)

    # Display the number of remaining columns
    print(f"\nNumber of Remaining Columns in Train Data: {len(remaining_columns)}")
    print(f"\nRemaining Columns in Test Data: {data.columns}")
    print(f"\nNumber of Data Points in Train Data: {len(data)}")
    print("\nRemoved Columns:")
    print(removed_columns)

    # Write cleaned data back to CSV
    # Replace 'cleaned_file.csv' with the desired name/path for your output file
    cleaned_file_path = your_path + "2.Cleaned/" + file_name + '_Cleaned.csv'
    data.to_csv(cleaned_file_path, index=False)
    return removed_columns

def compare_item(row,leg1,leg2):
    return 1 if row[leg1] == row[leg2] else 0

def clean_test(fp, r_col):
    
    try:
        # Get file in the folder
        files = os.listdir(fp)
        files = [f for f in files if f != '.DS_Store']
        file = files[0]
        file_name = files[0][:-4]
        
        # Check if there are any files
        if len(files) == 0:
            print("No files found in the folder.")

        else:
            file_path = os.path.join(fp, file)
            data = pd.read_csv(file_path, dtype=str)
            print(file)
            
    except Exception as e:
        print(f"An error occurred: {e}")
    
    #data = data.drop(columns=r_col+['disseminationTimestamp','sDRreceiptTimestamp','disseminationIdentifier'])
    
    data = data.drop(columns=r_col)

    # Convert any empty strings to NaN
    data = data.map(lambda x: None if x == '' else x)

    # Create interaction variable
    data['CurrencyIV'] = data.apply(lambda row: compare_item(row, 'leg1NotionalCurrency','leg2NotionalCurrency'), axis=1)
    data['NominalIV'] = data.apply(lambda row: compare_item(row, 'leg1NotionalAmount','leg2NotionalAmount'), axis=1)

    print(f"\nNumber of Remaining Columns in Test Data: {len(data.columns)}")
    #print(data.columns)
    print(f"\nNumber of Data Points in Test Data: {len(data)}")
    cleaned_file_path = your_path + "2.Cleaned/" + file_name + '_Cleaned.csv'
    data.to_csv(cleaned_file_path, index=False)
    
    

removed_cols = clean_train(train_folder_path)
clean_test(test_folder_path,removed_cols)
